cleanup and remove storage commands from the node api
