+++ cmd/pop/cli/cli.go
@@ -97,7 +97,6 @@ change until a first stable release. To get started run 'pop start'.
 			putCmd,
 			statusCmd,
 			commCmd,
-			storeCmd,
 			getCmd,
 			listCmd,
 			walletCmd,
